Reorganize java hadoop projects so we can share the schema class.
This just shifts things around so that the background processing task
and the elasticsearch import job are part of the same parent maven
module, so we can more easily share the Avro schema between them.
