+++ src/api-umbrella/hadoop-analytics/elasticsearch-import/.gitignore
