+++ src/api-umbrella/hadoop-analytics/elasticsearch-import/logging.properties
