+++ src/api-umbrella/hadoop-analytics/elasticsearch-import/src/main/java/apiumbrella/hadoop_analytics/OrcRowInspector.java
@@ -1,4 +1,4 @@
-package gov.nrel.apiumbrella;
+package apiumbrella.hadoop_analytics;
 
 import java.util.ArrayList;
 import java.util.List;
