+++ src/api-umbrella/hadoop-analytics/elasticsearch-import/src/main/resources/simplelogger.properties
@@ -1,0 +1,6 @@
+org.slf4j.simpleLogger.logFile=migrate.log
+org.slf4j.simpleLogger.defaultLogLevel=warn
+org.slf4j.simpleLogger.log.apiumbrella=info
+org.slf4j.simpleLogger.showDateTime=true
+org.slf4j.simpleLogger.dateTimeFormat=yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+org.slf4j.simpleLogger.showLogName=false
