+++ src/api-umbrella/hadoop-analytics/processor/.gitignore
@@ -1,6 +1,5 @@
 /.classpath
 /.project
 /.settings
-/bin/
 /dependency-reduced-pom.xml
 /target/
