+++ src/api-umbrella/hadoop-analytics/processor/src/main/java/apiumbrella/hadoop_analytics/ConvertLiveDataToOrc.java
