+++ src/api-umbrella/hadoop-analytics/processor/src/main/resources/simplelogger.properties
