+++ src/api-umbrella/hadoop-analytics/schema/.gitignore
@@ -1,6 +1,5 @@
 /.classpath
 /.project
 /.settings
-/bin/
 /dependency-reduced-pom.xml
 /target/
