+++ src/api-umbrella/hadoop-analytics/schema/src/main/resources/log.avsc
