--- src/api-umbrella/hadoop-analytics/src/main/resources/log.avsc
@@ -1,222 +1,0 @@
-{
-  "type": "record",
-  "name": "Log",
-  "fields": [
-    {
-      "name": "request_at",
-      "type": "long"
-    },
-    {
-      "name": "id",
-      "type": "string"
-    },
-    {
-      "name": "request_at_tz_offset",
-      "type": "int"
-    },
-    {
-      "name": "request_at_tz_year",
-      "type": "int"
-    },
-    {
-      "name": "request_at_tz_month",
-      "type": "int"
-    },
-    {
-      "name": "request_at_tz_week",
-      "type": "int"
-    },
-    {
-      "name": "request_at_tz_date",
-      "type": "string"
-    },
-    {
-      "name": "request_at_tz_hour",
-      "type": "int"
-    },
-    {
-      "name": "request_at_tz_minute",
-      "type": "int"
-    },
-    {
-      "name": "user_id",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "denied_reason",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_method",
-      "type": "string"
-    },
-    {
-      "name": "request_url_scheme",
-      "type": "string"
-    },
-    {
-      "name": "request_url_host",
-      "type": "string"
-    },
-    {
-      "name": "request_url_port",
-      "type": "int"
-    },
-    {
-      "name": "request_url_path",
-      "type": "string"
-    },
-    {
-      "name": "request_url_path_level1",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_path_level2",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_path_level3",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_path_level4",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_path_level5",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_path_level6",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_url_query",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_ip",
-      "type": "string"
-    },
-    {
-      "name": "request_ip_country",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_ip_region",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_ip_city",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_ip_lat",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "request_ip_lon",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "request_user_agent",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_user_agent_type",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_user_agent_family",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_size",
-      "type": ["null", "int"]
-    },
-    {
-      "name": "request_accept",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_accept_encoding",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_content_type",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_connection",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_origin",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_referer",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "request_basic_auth_username",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_status",
-      "type": "int"
-    },
-    {
-      "name": "response_content_type",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_content_length",
-      "type": ["null", "int"]
-    },
-    {
-      "name": "response_content_encoding",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_transfer_encoding",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_server",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_cache",
-      "type": ["null", "string"]
-    },
-    {
-      "name": "response_age",
-      "type": ["null", "int"]
-    },
-    {
-      "name": "response_size",
-      "type": ["null", "int"]
-    },
-    {
-      "name": "timer_response",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "timer_backend_response",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "timer_internal",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "timer_proxy_overhead",
-      "type": ["null", "double"]
-    },
-    {
-      "name": "log_imported",
-      "type": ["null", "boolean"]
-    }
-  ]
-}
